requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.0
fastapi-versioning==0.10.0
greenlet==3.1.0
h11==0.14.0
idna==3.10
inflect==7.4.0
more-itertools==10.5.0
passlib==1.7.4
psycopg2==2.9.9
py-automapper==1.2.3
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlacodegen-v2==0.1.4
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
typeguard==4.3.0
typing_extensions==4.12.2
uvicorn==0.30.6